recovery: removed all potential programming problems
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@25336 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 5c72ab025cb416c3a7e22e8cf3cb65f3a37e3e99

+++ server/recovery/de.uniol.inf.is.odysseus.recovery.incomingelements/src/de/uniol/inf/is/odysseus/recovery/incomingelements/sourcesync/physicaloperator/SourceRecoveryPO.java
@@ -98,7 +98,7 @@ public class SourceRecoveryPO<StreamObject extends IStreamObject<IMetaAttribute>
 		this.mNeedToAdjustOffset = false;
 		this.mRecoveryKafkaAccess = new KafkaConsumerAccess(this.mSourceAccess
 				.getAccessAOName().getResourceName(),
-				this.mRecoveryKafkaConsumer, this.mOffset);
+				this.mRecoveryKafkaConsumer, this.mOffset.longValue());
 		this.mRecoveryKafkaAccess.start();
 		super.process_open();
 	}
